used cached colors to set RGB at once when a single channel is changed; use validateRange() in command handler to ensure proper input values; try to catch the Hue=360 to convert it to Hue=0
